# Optimized support function for large meshes