@reduz 2022-08-14
This is very interesting. I was planning to do a different type of optimization for these kind of shapes which should hopefully be faster as its O(1).

My idea is to use [octahedral encoding](https://knarkowicz.wordpress.com/2014/04/16/octahedron-normal-vector-encoding/) for all possible directions. This way, a support function maps a normalized 2D unit vector to a 2D position in a 2D texture, then it only needs to check the 4 surrounding points found and one of them is going to be the correct result. There are some quirks in the implementation that are not so obvious if you are not familiar with type of encoding that I would need to look up if anyone is interested in doing.