@reduz 2022-08-14
In any case, if you are interested in physics engine discussion, I would be happy to invite you to the #physics channel of our rocketchat instance. https://chat.godotengine.org