@peastman 2022-08-14
Is your idea to encode the support function as an image?  I think that's equivalent to replacing the high resolution mesh with a simplified one whose vertices correspond to the pixels in the image?  Or you could think of it as applying a lowpass filter to the support function.  It should work well for smooth meshes, less well for ones with fine details.  Think of a long thin object like a spear.  If the tip of the spear doesn't exactly line up with a pixel, it will get shortened.

Another possibility is to combine the two approaches.  Use the image to get the initial guess for the vertex, then follow with the hill climbing stage to find the exact support point.  That would trade off more work at object creation time (depending on the size of the image) to reduce the number of iterations of hill climbing when a collision happens.

> In any case, if you are interested in physics engine discussion, I would be happy to invite you to the #physics channel of our rocketchat instance.

That would be great!