@peastman 2022-10-02
I went ahead and fixed the above in the more verbose way.  Let me know if it looks ok now.