@akien-mga 2022-10-12
CI is failing due to sign-compare warnings raised by both GCC and Clang:
```
servers/physics_3d/godot_shape_3d.cpp:827:19: error: comparison of integers of different signs: 'int' and 'unsigned int' [-Werror,-Wsign-compare]
        if (vertex_count > 3 * extreme_vertices.size()) {
            ~~~~~~~~~~~~ ^ ~~~~~~~~~~~~~~~~~~~~~~~~~~~
servers/physics_3d/godot_shape_3d.cpp:858:20: error: comparison of integers of different signs: 'int' and 'unsigned int' [-Werror,-Wsign-compare]
        for (int i = 0; i < extreme_vertices.size(); i++) {
                        ~ ^ ~~~~~~~~~~~~~~~~~~~~~~~
servers/physics_3d/godot_shape_3d.cpp:865:30: error: comparison of integers of different signs: 'unsigned int' and 'int' [-Werror,-Wsign-compare]
        if (extreme_vertices.size() == mesh.vertices.size()) {
            ~~~~~~~~~~~~~~~~~~~~~~~ ^  ~~~~~~~~~~~~~~~~~~~~
servers/physics_3d/godot_shape_3d.cpp:875:21: error: comparison of integers of different signs: 'int' and 'unsigned int' [-Werror,-Wsign-compare]
                for (int i = 0; i < vertex_neighbors[best_vertex].size(); i++) {
                                ~ ^ ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
servers/physics_3d/godot_shape_3d.cpp:1129:30: error: comparison of integers of different signs: 'unsigned int' and 'int' [-Werror,-Wsign-compare]
        if (extreme_vertices.size() < mesh.vertices.size()) {
            ~~~~~~~~~~~~~~~~~~~~~~~ ^ ~~~~~~~~~~~~~~~~~~~~
```