@peastman 2022-10-15
A test scene is attached.  It involves a collision between two object that each have a few hundred vertices.  It's unplayably slow either with or without this change, but much faster with the change than without.

Because of the way the SAT code works (it calls `project_range()` rather than `get_support()`), the new algorithm is only used for meshes with more than about 80 vertices.  I originally wrote this to be combined with #63702, which is support based and therefore provides more benefit even for smaller meshes.  I believe @reduz is currently planning to implement GJK instead, which is also support based so it should similarly benefit.

[gd4test4.zip.txt](https://github.com/godotengine/godot/files/9791043/gd4test4.zip.txt)