@akien-mga 2022-10-17
I ran the workflows to see if it passes CI but please note that this still requires to be squashed before it can be merged.
See [PR workflow](https://docs.godotengine.org/en/latest/community/contributing/pr_workflow.html#the-interactive-rebase) for instructions.