@peastman 2022-10-17
There was one more signed/unsigned comparison that needed fixing first.  I just made the change.  I can squash it once all tests are passing.

Github has a really useful feature that can do that automatically.  In the repository settings, go to the General tab and make sure "Allow squash merging" is checking.  Once it is, you can click on the arrow next to the "Merge" button for a PR and select "Squash and merge".  A benefit of doing it that way is that it generates a name for the squash commit that contains the name and number of the PR.  When you look through the commit history, every commit becomes a hyperlink to the corresponding PR.  Just something to consider for the future.