@akien-mga 2022-10-17
We don't use the squash-and-merge feature GitHub provides for various reasons. We prefer that PRs are squashed before merging so that the history we merge matches the one in the PR.