@akien-mga 2022-10-17
Yes, you can use `warnings=extra` (and `werror=yes` to treat as errors), or `dev_mode=yes` which would set both (+ `verbose=yes` and `tests=yes`).