@clayjohn 2022-10-27
Thanks and congratulations on your first merged contribution! I'm really forward to the physics optimizations you have in mind